Update various parts to run the server
- Change a lot of source code with comments
- Change config-related parts with comments
- Rename directory from /src to /pkg
- Add `pyproject.toml` to lint Python source code
- Update Dockerfile
