+++ .gitignore
@@ -20,11 +20,11 @@
 vendor/
 
 # Binaries
-src/src
-src/cm-beetle
-src/cm-beetle-arm
-src/cli/tbctl
-src/cli/cli
+pkg/src
+pkg/cm-beetle
+pkg/cm-beetle-arm
+pkg/cli/tbctl
+pkg/cli/cli
 
 # AutoGenerated DB data directory
 meta_db/
@@ -33,22 +33,22 @@ meta_db/
 container-volume/
 
 # Config file for test
-src/testclient/scripts/sequentialFullTest/executionStatus*
-src/testclient/scripts/sequentialFullTest/ansibleAutoConf/*
-src/testclient/scripts/credentials*
+pkg/testclient/scripts/sequentialFullTest/executionStatus*
+pkg/testclient/scripts/sequentialFullTest/ansibleAutoConf/*
+pkg/testclient/scripts/credentials*
 
 # Swagger docs for REST APIs
-src/api/rest/docs/docs.go
-src/api/rest/docs/swagger.json
-src/api/rest/docs/swagger.yaml
+pkg/api/rest/docs/docs.go
+pkg/api/rest/docs/swagger.json
+pkg/api/rest/docs/swagger.yaml
 
 # Config file for CM-Beetle
 conf/
 conf/credentials.conf
 
 # Runtime files
-src/benchmarking.csv
-src/rttmap.csv
+pkg/benchmarking.csv
+pkg/rttmap.csv
 
 # Files generated by IDE
 .idea/
@@ -59,3 +59,6 @@ log/
 
 # Files generated by Python virtual environment
 .venv/
+
+# Files generated by system_info.py
+system_info.yaml
