+++ pkg/.golangci.yml
