+++ pkg/api/rest/server/common/config.go
@@ -20,7 +20,7 @@ import (
 
 	"github.com/labstack/echo/v4"
 
-	"github.com/cloud-barista/cm-beetle/src/core/common"
+	"github.com/cloud-barista/cm-beetle/pkg/core/common"
 )
 
 // RestInitConfig godoc
