+++ pkg/core/common/utility.go
